+++ Changelog.md
@@ -4,6 +4,7 @@
 
  - Added `mutate(_:using:)` function to mutate values directly
  - `PartialRangeFrom`, `PartialRangeThrough`, and `PartialRangeUpTo` now conform to `Equatable` `Hashable`, `CustomStringConvertible`, `CustomDebugStringConvertible`, and `CustomReflectable`
+ - Added `AnyEquatable`, a type-erased wrapper for `Equatable`
 
 ## 2.0.0